testing

+++ src/main/java/com/voisovych/internetmarket/servis/ItemCRUDService.java
@@ -2,15 +2,26 @@ package com.voisovych.internetmarket.servis;
 
 import com.voisovych.internetmarket.model.Item;
 import com.voisovych.internetmarket.repository.ItemCRUDRepository;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import javax.persistence.EntityExistsException;
+import javax.transaction.Transactional;
 import java.util.List;
 
 @Service
 public class ItemCRUDService {
 
+    private ItemCRUDRepository itemCRUDRepository;
+
     @Autowired
-    ItemCRUDRepository itemCRUDRepository;
+    public ItemCRUDService(ItemCRUDRepository itemCRUDRepository) {
+        this.itemCRUDRepository = itemCRUDRepository;
+    }
+
+    Logger logger = LogManager.getLogger(ItemCRUDService.class);
 
     public List<Item> search(Item item) {
         return itemCRUDRepository.findAllByCountAndNameAndDescriptionAndNumberAndPriceAndTypeAndCreationDate (
@@ -26,16 +37,27 @@ public class ItemCRUDService {
         return itemCRUDRepository.findAll();
     }
 
-    public void save(Item item){
-        itemCRUDRepository.save(item);
+
+    public Item save(Item item) {
+
+        return itemCRUDRepository.save(item);
     }
 
     public Item findById(Long id){
         return itemCRUDRepository.findAllById(id);
     }
 
-    public void delete (Long id){
-        itemCRUDRepository.deleteById(id);
+    public boolean delete (Long id){
+        if(itemCRUDRepository.findById(id).isPresent()) {
+            itemCRUDRepository.deleteById(id);
+            return true;
+        }
+        throw new EntityExistsException("Not found");
+    }
+
+    @Transactional
+    public void saveAll(Iterable<Item> iterable) {
+        itemCRUDRepository.saveAll(iterable);
     }
 
 //    public List<Item> searchByTypeName(String type, String name){
